Fix clazy warnings for MeteorShowers plugin

+++ plugins/MeteorShowers/src/MeteorShowersMgr.cpp
@@ -361,8 +361,8 @@ void MeteorShowersMgr::updateDownloadProgress(qint64 bytesReceived, qint64 bytes
 		//Round to the greatest possible derived unit
 		while (bytesTotal > 1024)
 		{
-			bytesReceived = static_cast<qint64>(std::floor(bytesReceived / 1024.));
-			bytesTotal    = static_cast<qint64>(std::floor(bytesTotal / 1024.));
+			bytesReceived = static_cast<qint64>(std::floor(static_cast<double>(bytesReceived) / 1024.));
+			bytesTotal    = static_cast<qint64>(std::floor(static_cast<double>(bytesTotal)    / 1024.));
 		}
 		currentValue = static_cast<int>(bytesReceived);
 		endValue = static_cast<int>(bytesTotal);
